# Store rawValue as a string instead of []bytes