@ianlopshire 2019-10-06
@alex, looks good!

Can you resolve the merge conflict? Then I will merge and cut a new release.

Thanks for submitting!