@alex 2019-10-06
One note: I'd expect this to get even faster once my change to `bufio.Reader.ReadString` is merged into Go itself: https://go-review.googlesource.com/c/go/+/199257